Implement users default discount rate to order (ADNAPI-158)

+++ src/Adanom.Ecommerce.API.Commands/Commands/Identity/User/Models/UpdateUserDefaultDiscountRateRequest.cs
@@ -4,6 +4,6 @@ namespace Adanom.Ecommerce.API.Commands.Models
     {
         public Guid Id { get; set; }
 
-        public byte DefaultDiscountRate { get; set; }
+        public decimal DefaultDiscountRate { get; set; }
     }
 }
